+++ src/coq_elpi_HOAS.ml
@@ -2318,6 +2318,21 @@ let reachable sigma roots acc =
       prlist_with_sep spc Evar.print (Evar.Set.elements res));
   res
 
+let solution2evd sigma0 { API.Data.constraints; assignments; state; pp_ctx } roots =
+  let state, solved_goals, _, _gls = elpi_solution_to_coq_solution constraints state in
+  let sigma = get_sigma state in
+  let all_goals = reachable sigma roots Evar.Set.empty in
+  let declared_goals, shelved_goals =
+    get_declared_goals (Evar.Set.diff all_goals solved_goals) constraints state assignments pp_ctx in
+  debug Pp.(fun () -> str "Goals: " ++ prlist_with_sep spc Evar.print declared_goals);
+  debug Pp.(fun () -> str "Shelved Goals: " ++ prlist_with_sep spc Evar.print shelved_goals);
+  Evd.fold_undefined (fun k _ sigma ->
+    if Evar.Set.mem k all_goals || Evd.mem sigma0 k then sigma
+    else Evd.remove sigma k
+    ) sigma sigma,
+  declared_goals,
+  shelved_goals
+
 let tclSOLUTION2EVD sigma0 { API.Data.constraints; assignments; state; pp_ctx } =
   let open Proofview.Unsafe in
   let open Tacticals in